+++ src/main/gen/org/c3lang/intellij/psi/C3NamedIdent.java
@@ -5,6 +5,6 @@ import java.util.List;
 import org.jetbrains.annotations.*;
 import com.intellij.psi.PsiElement;
 
-public interface C3CtConcat extends PsiElement {
+public interface C3NamedIdent extends PsiElement {
 
 }