+++ src/main/java/org/c3lang/intellij/C3TokenSets.java
@@ -87,8 +87,6 @@ public interface C3TokenSets
     TokenSet CT_KEYWORDS = TokenSet.create(C3Types.KW_CT_ALIGNOF,
                                            C3Types.KW_CT_ASSERT,
                                            C3Types.KW_CT_CASE,
-                                           C3Types.KW_CT_AND,
-                                           C3Types.KW_CT_OR,
                                            C3Types.KW_CT_IS_CONST,
                                            C3Types.KW_CT_DEFAULT,
                                            C3Types.KW_CT_DEFINED,